Make all previous scripting runnable from main.py
Previous scripts (i.e. serve and migrations) are now run through main.py
where serve is default and migrations is run when -m is used as a cmd
line arg.

Also adds --production as a cmd line arg, but it's not used yet

Migrations was previously broken because running it as its own project
meant it couldn't actually access the config_utils. Running it this way
fixes that issue.